fix: 쪽마다 새로 시작하는 구역에선 각주 번호를 미리 받지 않는다 (#95)
번호를 문서 순서로 되돌리며 앞 조각이 컨테이너를 걸어 번호까지 확정하게 했는데,
"쪽마다 새로 시작"(표 134 numberingMode 2)에서는 그 번호가 낡는다 — run 사이
cacheCurrentPage가 카운터를 시작 번호로 되돌리므로, 미리 받은 번호가 flush되는
쪽의 첫 직접 각주와 겹쳐 같은 쪽에 1)이 둘 생긴다.

그 모드에서 번호는 정의상 **그려질 쪽**의 함수고 쪽마다 1로 돌아가 보존할 순서
관계도 없으므로, 미루지 않고 마지막 조각이 걷게 둔다 (번호가 제 쪽 카운터에서
나온다). 리뷰가 제안한 flush 시점 rebase 대신 이 게이트를 쓴 이유는 "리셋이
있었나"를 알기 위한 번호 세대 상태가 필요 없어서다.

가드가 클래스 본문을 400줄(SwiftLint error 임계) 위로 밀어 컨테이너/중첩 수집
규약 5종을 HwpFootnoteContainerCollectionTests로 옮겼다 — 조각 귀속(실제
paginator)과 수집기 규약(단위)이 원래 다른 축이라 파일도 그 축으로 가른다.

실측: 새 가드는 게이트 없이 [1, 1]로 실패하고 게이트 후 [1, 2]다. 렌더 해시는
양 폰트 모드 전 픽스처 무변화, 커밋된 기준선 4종 포함 1,195 테스트 통과,
SwiftLint error 0, 추적 파일 swiftformat 0/558.
