+++ Tests/HwpKitCoreTests/HwpFootnoteFragmentAttributionTests.swift
@@ -204,258 +204,6 @@ import XCTest
             expect(self.bodyText(on: second)).toNot(contain("10)"))
         }
 
-        /// 컨테이너 (글상자·도형) 는 `appendControlBlocks`가 모든 조각을 놓은 **뒤**
-        /// 방출해 마지막 조각 페이지에 그려진다. 그 안의 각주를 앞 조각에서 걷으면
-        /// 각주와 그것을 그리는 컨테이너가 갈린다 — 앞 조각에서는 내려가지 않고,
-        /// 마지막 조각에서는 서수 범위 **밖**이어도 내려가야 한다 (안 그러면 유실).
-        func testNestedNotesWaitForTheFragmentThatDrawsTheirContainer() {
-            var coordinator = HwpFootnoteCoordinator(
-                index: HwpIndex(from: CoreHwp.HwpFile()),
-                fontResolver: .testDeterministic
-            )
-            var host = CoreHwp.HwpParagraph()
-            host.ctrlHeaderArray = [
-                .header(HwpSynthetic.listControl(ctrlId: .header, paragraphs: [])),
-            ]
-            var nested = CoreHwp.HwpParagraph()
-            nested.ctrlHeaderArray = [
-                .footnote(HwpSynthetic.listControl(
-                    ctrlId: .footnote,
-                    paragraphs: [HwpSynthetic.noteParagraph(
-                        " 컨테이너 안 각주",
-                        autoNumber: HwpSynthetic.autoNumberControl(kind: 1, decorationTail: ")")
-                    )]
-                )),
-            ]
-            // 컨테이너만 자식을 갖는다 — 모든 컨트롤에 돌려주면 각주 리스트가
-            // 자기 자신을 자식으로 받아 깊이 상한까지 재귀한다.
-            let children: HwpFootnoteCoordinator.ChildParagraphs = { ctrl in
-                if case .header = ctrl {
-                    return [(nested, .textbox)]
-                }
-                return []
-            }
-            let environment = HwpFootnoteCoordinator.Environment(
-                contentWidth: 400, footnoteShape: nil
-            )
-
-            coordinator.collectFootnotes(
-                from: host,
-                includeTableCells: false,
-                ordinals: 0 ..< 1,
-                collectsNested: false,
-                environment: environment,
-                childParagraphs: children
-            )
-            expect(coordinator.pendingFootnotes).to(beEmpty())
-            // 미룬 각주는 이 페이지 몫이 아니므로 예약도 하지 않는다 — 예약이
-            // 늘면 effectiveContentHeight가 줄어 본문 절단점이 흔들린다.
-            expect(coordinator.footnoteReservedHeight) == 0
-
-            coordinator.collectFootnotes(
-                from: host,
-                includeTableCells: false,
-                ordinals: 1 ..< 1,
-                collectsNested: true,
-                environment: environment,
-                childParagraphs: children
-            )
-            expect(coordinator.pendingFootnotes.count) == 1
-            expect(coordinator.footnoteReservedHeight) > 0
-            // 버퍼를 풀고 **또** 걷지 않는다 — 그러면 같은 각주가 두 번 실리고
-            // 번호도 하나 더 소비된다.
-            expect(coordinator.footnoteCounter) == 1
-        }
-
-        /// 번호는 **문서 순서**다 — 컨테이너 안 각주가 그 뒤의 직접 각주보다 앞선다.
-        /// 배치를 미루면서 번호까지 미루면 뒤의 직접 각주가 먼저 카운터를 가져가
-        /// 둘이 뒤바뀐다 (#95 리뷰 r3708636335).
-        func testDeferredContainerNotesKeepTheirDocumentOrderNumbers() {
-            var coordinator = HwpFootnoteCoordinator(
-                index: HwpIndex(from: CoreHwp.HwpFile()),
-                fontResolver: .testDeterministic
-            )
-            var host = CoreHwp.HwpParagraph()
-            host.ctrlHeaderArray = [
-                .header(HwpSynthetic.listControl(ctrlId: .header, paragraphs: [])),
-                .footnote(HwpSynthetic.listControl(
-                    ctrlId: .footnote,
-                    paragraphs: [HwpSynthetic.noteParagraph(
-                        " 뒤따르는 직접 각주",
-                        autoNumber: HwpSynthetic.autoNumberControl(kind: 1, decorationTail: ")")
-                    )]
-                )),
-            ]
-            var nested = CoreHwp.HwpParagraph()
-            nested.ctrlHeaderArray = [
-                .footnote(HwpSynthetic.listControl(
-                    ctrlId: .footnote,
-                    paragraphs: [HwpSynthetic.noteParagraph(
-                        " 컨테이너 안 각주",
-                        autoNumber: HwpSynthetic.autoNumberControl(kind: 1, decorationTail: ")")
-                    )]
-                )),
-            ]
-            let children: HwpFootnoteCoordinator.ChildParagraphs = { ctrl in
-                if case .header = ctrl {
-                    return [(nested, .textbox)]
-                }
-                return []
-            }
-            let environment = HwpFootnoteCoordinator.Environment(
-                contentWidth: 400, footnoteShape: nil
-            )
-
-            // 앞 조각: 컨테이너(서수 0)와 직접 각주(서수 1)가 같은 조각에 있다
-            coordinator.collectFootnotes(
-                from: host,
-                includeTableCells: false,
-                ordinals: 0 ..< 2,
-                collectsNested: false,
-                environment: environment,
-                childParagraphs: children
-            )
-            expect(coordinator.pendingFootnotes.count) == 1
-            let directNumber = coordinator.pendingFootnotes[0].number
-
-            coordinator.collectFootnotes(
-                from: host,
-                includeTableCells: false,
-                ordinals: 2 ..< 2,
-                collectsNested: true,
-                environment: environment,
-                childParagraphs: children
-            )
-            expect(coordinator.pendingFootnotes.count) == 2
-            let nestedNumber = coordinator.pendingFootnotes[1].number
-
-            expect(nestedNumber) == 0
-            expect(directNumber) == 1
-        }
-
-        /// 각주는 **이 조각이 곧바로 배치**하므로 그 안쪽 노트도 같은 조각이 걷는다
-        /// — 미루면 참조와 갈리고 번호 순서가 밀린다 (#95 리뷰). 마지막 조각이 또
-        /// 걷어서도 안 된다 (같은 노트를 두 번 셈).
-        func testNotesInsideACollectedFootnoteAreCollectedWithIt() {
-            var coordinator = HwpFootnoteCoordinator(
-                index: HwpIndex(from: CoreHwp.HwpFile()),
-                fontResolver: .testDeterministic
-            )
-            var host = CoreHwp.HwpParagraph()
-            host.ctrlHeaderArray = [
-                .footnote(HwpSynthetic.listControl(
-                    ctrlId: .footnote,
-                    paragraphs: [HwpSynthetic.noteParagraph(
-                        " 바깥 각주",
-                        autoNumber: HwpSynthetic.autoNumberControl(kind: 1, decorationTail: ")")
-                    )]
-                )),
-            ]
-            // 각주 문단 **안**의 노트 — 미주로 두어 스텁 재귀가 멈추게 한다
-            var inner = CoreHwp.HwpParagraph()
-            inner.ctrlHeaderArray = [
-                .endnote(HwpSynthetic.listControl(
-                    ctrlId: .endnote,
-                    paragraphs: [HwpSynthetic.noteParagraph(
-                        " 각주 안 미주",
-                        autoNumber: HwpSynthetic.autoNumberControl(kind: 1, decorationTail: ")")
-                    )]
-                )),
-            ]
-            let children: HwpFootnoteCoordinator.ChildParagraphs = { ctrl in
-                if case .footnote = ctrl {
-                    return [(inner, .footnote)]
-                }
-                return []
-            }
-            let environment = HwpFootnoteCoordinator.Environment(
-                contentWidth: 400, footnoteShape: nil
-            )
-
-            coordinator.collectFootnotes(
-                from: host,
-                includeTableCells: false,
-                ordinals: 0 ..< 1,
-                collectsNested: false,
-                environment: environment,
-                childParagraphs: children
-            )
-            expect(coordinator.pendingFootnotes.count) == 1
-            expect(coordinator.pendingEndnotes.count) == 1
-
-            coordinator.collectFootnotes(
-                from: host,
-                includeTableCells: false,
-                ordinals: 1 ..< 1,
-                collectsNested: true,
-                environment: environment,
-                childParagraphs: children
-            )
-            expect(coordinator.pendingFootnotes.count) == 1
-            expect(coordinator.pendingEndnotes.count) == 1
-        }
-
-        /// 미주는 문서·구역 끝에서 흐름 배치되므로 (`placeFlow`) 이 조각이 그리지
-        /// 않는다 — 안쪽 각주를 지금 걷으면 참조는 문서 끝에, 각주는 본문 쪽에
-        /// 남는다. 글상자·도형과 같이 마지막 조각으로 미룬다 (#95 리뷰).
-        func testEndnoteDescendantsWaitForTheFinalFragment() {
-            var coordinator = HwpFootnoteCoordinator(
-                index: HwpIndex(from: CoreHwp.HwpFile()),
-                fontResolver: .testDeterministic
-            )
-            var host = CoreHwp.HwpParagraph()
-            host.ctrlHeaderArray = [
-                .endnote(HwpSynthetic.listControl(
-                    ctrlId: .endnote,
-                    paragraphs: [HwpSynthetic.noteParagraph(
-                        " 미주",
-                        autoNumber: HwpSynthetic.autoNumberControl(kind: 1, decorationTail: ")")
-                    )]
-                )),
-            ]
-            var inner = CoreHwp.HwpParagraph()
-            inner.ctrlHeaderArray = [
-                .footnote(HwpSynthetic.listControl(
-                    ctrlId: .footnote,
-                    paragraphs: [HwpSynthetic.noteParagraph(
-                        " 미주 안 각주",
-                        autoNumber: HwpSynthetic.autoNumberControl(kind: 1, decorationTail: ")")
-                    )]
-                )),
-            ]
-            let children: HwpFootnoteCoordinator.ChildParagraphs = { ctrl in
-                if case .endnote = ctrl {
-                    return [(inner, .footnote)]
-                }
-                return []
-            }
-            let environment = HwpFootnoteCoordinator.Environment(
-                contentWidth: 400, footnoteShape: nil
-            )
-
-            coordinator.collectFootnotes(
-                from: host,
-                includeTableCells: false,
-                ordinals: 0 ..< 1,
-                collectsNested: false,
-                environment: environment,
-                childParagraphs: children
-            )
-            expect(coordinator.pendingEndnotes.count) == 1
-            expect(coordinator.pendingFootnotes).to(beEmpty())
-
-            coordinator.collectFootnotes(
-                from: host,
-                includeTableCells: false,
-                ordinals: 1 ..< 1,
-                collectsNested: true,
-                environment: environment,
-                childParagraphs: children
-            )
-            expect(coordinator.pendingFootnotes.count) == 1
-            expect(coordinator.pendingEndnotes.count) == 1
-        }
-
         // MARK: - 조각별 컨트롤 서수 분할
 
         /// 그 조각에 그려진 마커만 담은 텍스트 — 배치가 낸 슬라이스의 최소 형상.
